# docs/swagger.yaml
definitions:
  db.AddBackfillCrawlerParams:
    properties:
      chainId:
        type: integer
      collectionAddress:
        type: string
      currentBlock:
        type: integer
    type: object
  db.AddChainParams:
    properties:
      blockTime:
        type: integer
      chain:
        type: string
      chainId:
        type: integer
      explorer:
        type: string
      id:
        type: integer
      latestBlock:
        type: integer
      name:
        type: string
      rpcUrl:
        type: string
    type: object
  db.AssetType:
    enum:
    - ERC721
    - ERC1155
    - ERC20
    - ORDER
    type: string
    x-enum-varnames:
    - AssetTypeERC721
    - AssetTypeERC1155
    - AssetTypeERC20
    - AssetTypeORDER
  response.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
  response.ResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  utils.AddNewAssetParamsSwagger:
    properties:
      chainID:
        type: integer
      collectionAddress:
        type: string
      decimalData:
        type: integer
      initialBlock:
        type: integer
      lastUpdated:
        type: string
      type:
        $ref: '#/definitions/db.AssetType'
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8085
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /backfill:
    post:
      consumes:
      - application/json
      description: Add a new asset collection to the chain
      parameters:
      - description: Asset collection information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.AddBackfillCrawlerParams'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Add a new asset collection to the chain
      tags:
      - asset
      - backfill
  /chain:
    get:
      consumes:
      - application/json
      description: Get all supported chains
      parameters:
      - description: Chain Id
        in: query
        name: chain_id
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all supported chains
      tags:
      - chains
    post:
      consumes:
      - application/json
      description: Add a new chain
      parameters:
      - description: Add a new chain
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.AddChainParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a new chain
      tags:
      - chains
  /chain/{chain_id}/collection:
    get:
      consumes:
      - application/json
      description: Retrieve all asset collections associated with the specified chain
        Id.
      parameters:
      - description: Chain Id
        in: path
        name: chain_id
        required: true
        type: string
      - description: Collection Address
        in: query
        name: collection_address
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all asset in a collection of the chain
      tags:
      - asset
    post:
      consumes:
      - application/json
      description: Add a new asset collection to the chain
      parameters:
      - description: Chain Id
        in: path
        name: chain_id
        required: true
        type: string
      - description: Asset collection information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.AddNewAssetParamsSwagger'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Add a new asset collection to the chain
      tags:
      - asset
  /chain/{chain_id}/collection/{collection_address}/{token_id}:
    get:
      consumes:
      - application/json
      description: Get detail of asset collection of the chain
      parameters:
      - description: Chain Id
        in: path
        name: chain_id
        required: true
        type: integer
      - description: Collection Address
        in: path
        name: collection_address
        required: true
        type: string
      - description: Token Id
        in: path
        name: token_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get detail of asset collection of the chain
      tags:
      - asset
  /chain/{chain_id}/collection/{collection_address}/assets:
    get:
      consumes:
      - application/json
      description: Get all asset collection of the chain
      parameters:
      - description: Chain Id
        in: path
        name: chain_id
        required: true
        type: integer
      - description: Collection Address
        in: path
        name: collection_address
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - collectionFormat: multi
        description: Token Ids
        in: query
        items:
          type: string
        name: token_id
        type: array
      - description: Owner Address
        in: query
        name: owner
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all asset collection of the chain
      tags:
      - asset
  /chain/{chain_id}/nft-assets:
    get:
      consumes:
      - application/json
      description: Get all NFT asset collection of the chain
      parameters:
      - description: Chain Id
        in: path
        name: chain_id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all NFT asset collection of the chain
      tags:
      - asset
  /chain/order:
    get:
      consumes:
      - application/json
      description: Get all Exchange Order records of the chain
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Chain id
        in: query
        name: chain_id
        type: string
      - description: Maker Address
        in: query
        name: maker
        type: string
      - description: Taker Address
        in: query
        name: taker
        type: string
      - description: Signature
        in: query
        name: sig
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Transaction Hash
        in: query
        name: tx_hash
        type: string
      - description: Asset Id
        in: query
        name: asset_id
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all Exchange Order records of the chain
      tags:
      - asset
  /history:
    get:
      consumes:
      - application/json
      description: Get History of a transaction
      parameters:
      - description: Tx Hash
        in: query
        name: tx_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Get History of a transaction
      tags:
      - history
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"